backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9
Pillow==10.4.0
pydantic-settings==2.4.0
